refactor: split rest template import modal
